script for generating speculator training set on nersc

+++ run/speculator_training.py
@@ -1,4 +1,4 @@
-import os
+import os, sys
 import numpy as np 
 # -- gqp_mc --
 from gqp_mc import util as UT
@@ -27,7 +27,7 @@ def sample_prior(n_sample):
     return prior_min + (prior_max - prior_min) * np.random.uniform(size=(n_sample, len(prior_min)))
 
 
-def train_desi_simpledust(seed=0): 
+def train_desi_simpledust(ibatch, seed=0): 
     ''' generate FSPS training set for DESI with simple calzetti dust model.
     '''
     # set random seed
@@ -35,6 +35,8 @@ def train_desi_simpledust(seed=0):
 
     batch = 100
     nspec = 6000
+
+    assert ibatch < batch
     
     # wavelength range set to cover the DESI spectra and photometric filter
     # wavelength range 
@@ -44,36 +46,41 @@ def train_desi_simpledust(seed=0):
 
     theta_train = sample_prior(batch*nspec)
     theta_train[10000:,:4] =\
-            speculate._transform_to_SFH_basis(np.random.uniform(size=(batch*nspec,4)))
+            speculate._transform_to_SFH_basis(np.random.uniform(size=(batch*nspec-10000,4)))
     
-    # save FSPS wavelength 
     w_fsps, _ = speculate._fsps_model(theta_train[0])
     wlim = (w_fsps >= wmin) & (w_fsps <= wmax)
 
     fwave = os.path.join(UT.dat_dir(), 'speculator', 'wave_fsps.npy')
-    np.save(fwave, w_fsps[wlim])
+    if not os.path.isfile(fwave):
+        # save FSPS wavelength if not saved 
+        np.save(fwave, w_fsps[wlim])
 
-    for i in range(batch):
-        ftheta = os.path.join(UT.dat_dir(), 'speculator', 'DESI_simpledust.theta_train.%i.npy' % i)
-        fspectrum = os.path.join(UT.dat_dir(), 'speculator', 'DESI_simpledust.logspectrum_fsps_train.%i.npy' % i)
-    
-        if os.path.isfile(ftheta) and os.path.isfile(fspectrum): 
-            pass
-        else: 
-            print('--- batch %i ---' % i)
+    ftheta = os.path.join(UT.dat_dir(), 'speculator',
+            'DESI_simpledust.theta_train.%i.seed%i.npy' % (ibatch, seed))
+    fspectrum = os.path.join(UT.dat_dir(), 'speculator',
+            'DESI_simpledust.logspectrum_fsps_train.%i.seed%i.npy' % 
+            (ibatch, seed))
 
-            i_batch = range(nspec*i,nspec*(i+1))
+    if os.path.isfile(ftheta) and os.path.isfile(fspectrum): 
+        print('--- batch %i already exists ---' % ibatch)
+        return None 
+    else: 
+        print('--- batch %i ---' % ibatch)
 
-            logspectra_train = []
-            for _theta in theta_train[i_batch]:
-                _, _spectrum = speculate._fsps_model(_theta)
-                logspectra_train.append(np.log(_spectrum[wlim]))
+        i_batch = range(nspec*ibatch,nspec*(ibatch+1))
 
-            np.save(ftheta, theta_train[i_batch])
-            np.save(fspectrum, np.array(logspectra_train))
+        logspectra_train = []
+        for _theta in theta_train[i_batch]:
+            _, _spectrum = speculate._fsps_model(_theta)
+            logspectra_train.append(np.log(_spectrum[wlim]))
+
+        np.save(ftheta, theta_train[i_batch])
+        np.save(fspectrum, np.array(logspectra_train))
 
     return None 
 
 
 if __name__=='__main__': 
-    train_desi_simpledust(seed=0)
+    ibatch = int(sys.argv[1])
+    train_desi_simpledust(ibatch, seed=0)
